Implement password recovery for forgotten password. Change some english text to hungarian in password recovery views.

+++ nptk/Models/IdentityModels.cs
@@ -81,12 +81,12 @@ namespace nptk.Models
 
         public DbSet<SignUp> SignUps { get; set; }
 
-        public decimal DistanceCount(int? Id)
+        public double DistanceCount(int? Id)
         {
-            decimal distanceCount = (from t in this.Tours
+            double distanceCount = (from t in this.Tours
                              from s in t.SignUps
                              where s.UserID == Id
-                             select (decimal?)t.Distance).Sum() ?? 0;
+                             select (double?)t.Distance).Sum() ?? 0;
             Debug.WriteLine("DistanceCount: " + distanceCount + " id: " + Id);
             return distanceCount;
         }
